+++ packages/google_maps_flutter/google_maps_flutter_android/android/src/main/java/io/flutter/plugins/googlemaps/Convert.java
@@ -4,9 +4,11 @@
 
 package io.flutter.plugins.googlemaps;
 
+import android.content.res.AssetManager;
 import android.graphics.Bitmap;
 import android.graphics.BitmapFactory;
 import android.graphics.Point;
+import androidx.annotation.NonNull;
 import androidx.annotation.VisibleForTesting;
 import com.google.android.gms.maps.CameraUpdate;
 import com.google.android.gms.maps.CameraUpdateFactory;
@@ -26,6 +28,9 @@ import com.google.android.gms.maps.model.RoundCap;
 import com.google.android.gms.maps.model.SquareCap;
 import com.google.android.gms.maps.model.Tile;
 import com.google.maps.android.clustering.Cluster;
+import io.flutter.FlutterInjector;
+import java.io.IOException;
+import java.io.InputStream;
 import com.google.maps.android.heatmaps.Gradient;
 import com.google.maps.android.heatmaps.WeightedLatLng;
 import com.google.maps.android.projection.SphericalMercatorProjection;
@@ -34,7 +39,6 @@ import java.util.Arrays;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.Set;
 
 /** Conversions between JSON-like values and GoogleMaps data types. */
 class Convert {
@@ -52,43 +56,67 @@ class Convert {
   public static final String HEATMAP_GRADIENT_START_POINTS_KEY = "startPoints";
   public static final String HEATMAP_GRADIENT_COLOR_MAP_SIZE_KEY = "colorMapSize";
 
-  // TODO(hamdikahloun): FlutterMain has been deprecated and should be replaced with FlutterLoader
-  //  when it's available in Stable channel: https://github.com/flutter/flutter/issues/70923.
-  @SuppressWarnings("deprecation")
-  private static BitmapDescriptor toBitmapDescriptor(Object o) {
+  private static BitmapDescriptor toBitmapDescriptor(
+      Object o, AssetManager assetManager, float density) {
     final List<?> data = toList(o);
-    switch (toString(data.get(0))) {
+    final String descriptorType = toString(data.get(0));
+    switch (descriptorType) {
       case "defaultMarker":
         if (data.size() == 1) {
           return BitmapDescriptorFactory.defaultMarker();
         } else {
-          return BitmapDescriptorFactory.defaultMarker(toFloat(data.get(1)));
+          final float hue = toFloat(data.get(1));
+          return BitmapDescriptorFactory.defaultMarker(hue);
         }
       case "fromAsset":
+        final String assetPath = toString(data.get(1));
         if (data.size() == 2) {
           return BitmapDescriptorFactory.fromAsset(
-              io.flutter.view.FlutterMain.getLookupKeyForAsset(toString(data.get(1))));
+              FlutterInjector.instance().flutterLoader().getLookupKeyForAsset(assetPath));
         } else {
+          final String assetPackage = toString(data.get(2));
           return BitmapDescriptorFactory.fromAsset(
-              io.flutter.view.FlutterMain.getLookupKeyForAsset(
-                  toString(data.get(1)), toString(data.get(2))));
+              FlutterInjector.instance()
+                  .flutterLoader()
+                  .getLookupKeyForAsset(assetPath, assetPackage));
         }
       case "fromAssetImage":
+        final String assetImagePath = toString(data.get(1));
         if (data.size() == 3) {
           return BitmapDescriptorFactory.fromAsset(
-              io.flutter.view.FlutterMain.getLookupKeyForAsset(toString(data.get(1))));
+              FlutterInjector.instance().flutterLoader().getLookupKeyForAsset(assetImagePath));
         } else {
           throw new IllegalArgumentException(
               "'fromAssetImage' Expected exactly 3 arguments, got: " + data.size());
         }
       case "fromBytes":
-        return getBitmapFromBytes(data);
+        return getBitmapFromBytesLegacy(data);
+      case "asset":
+        if (!(data.get(1) instanceof Map)) {
+          throw new IllegalArgumentException("'asset' expected a map as the second parameter");
+        }
+        final Map<?, ?> assetData = toMap(data.get(1));
+        return getBitmapFromAsset(
+            assetData,
+            assetManager,
+            density,
+            new BitmapDescriptorFactoryWrapper(),
+            new FlutterInjectorWrapper());
+      case "bytes":
+        if (!(data.get(1) instanceof Map)) {
+          throw new IllegalArgumentException("'bytes' expected a map as the second parameter");
+        }
+        final Map<?, ?> byteData = toMap(data.get(1));
+        return getBitmapFromBytes(byteData, density, new BitmapDescriptorFactoryWrapper());
       default:
         throw new IllegalArgumentException("Cannot interpret " + o + " as BitmapDescriptor");
     }
   }
 
-  private static BitmapDescriptor getBitmapFromBytes(List<?> data) {
+  // Used for deprecated fromBytes bitmap descriptor.
+  // Can be removed after support for "fromBytes" bitmap descriptor type is
+  // removed.
+  private static BitmapDescriptor getBitmapFromBytesLegacy(List<?> data) {
     if (data.size() == 2) {
       try {
         Bitmap bitmap = toBitmap(data.get(1));
@@ -101,6 +129,185 @@ class Convert {
           "fromBytes should have exactly one argument, interpretTileOverlayOptions the bytes. Got: "
               + data.size());
     }
+  }
+
+  /**
+   * Creates a BitmapDescriptor object from bytes data.
+   *
+   * <p>This method requires the `byteData` map to contain specific keys: 'byteData' for image
+   * bytes, 'bitmapScaling' for scaling mode, and 'imagePixelRatio' for scale ratio. It may
+   * optionally include 'width' and/or 'height' for explicit image dimensions.
+   *
+   * @param byteData a map containing the byte data and scaling instructions. Expected keys are:
+   *     'byteData': the actual bytes of the image, 'bitmapScaling': the scaling mode, either 'auto'
+   *     or 'none', 'imagePixelRatio': used with 'auto' bitmapScaling if width or height are not
+   *     provided, 'width' (optional): the desired width, which affects scaling if 'height' is not
+   *     provided, 'height' (optional): the desired height, which affects scaling if 'width' is not
+   *     provided
+   * @param density the density of the display, used to calculate pixel dimensions.
+   * @param bitmapDescriptorFactory is an instance of the BitmapDescriptorFactoryWrapper.
+   * @return BitmapDescriptor object from bytes data.
+   * @throws IllegalArgumentException if any required keys are missing in `byteData` or if the byte
+   *     data cannot be interpreted as a valid image.
+   */
+  @VisibleForTesting
+  public static BitmapDescriptor getBitmapFromBytes(
+      Map<?, ?> byteData, float density, BitmapDescriptorFactoryWrapper bitmapDescriptorFactory) {
+
+    final String byteDataKey = "byteData";
+    final String bitmapScalingKey = "bitmapScaling";
+    final String imagePixelRatioKey = "imagePixelRatio";
+
+    if (!byteData.containsKey(byteDataKey)) {
+      throw new IllegalArgumentException("'bytes' requires '" + byteDataKey + "' key.");
+    }
+    if (!byteData.containsKey(bitmapScalingKey)) {
+      throw new IllegalArgumentException("'bytes' requires '" + bitmapScalingKey + "' key.");
+    }
+    if (!byteData.containsKey(imagePixelRatioKey)) {
+      throw new IllegalArgumentException("'bytes' requires '" + imagePixelRatioKey + "' key.");
+    }
+
+    try {
+      Bitmap bitmap = toBitmap(byteData.get(byteDataKey));
+      String scalingMode = toString(byteData.get(bitmapScalingKey));
+      switch (scalingMode) {
+        case "auto":
+          final String widthKey = "width";
+          final String heightKey = "height";
+
+          final Double width =
+              byteData.containsKey(widthKey) ? toDouble(byteData.get(widthKey)) : null;
+          final Double height =
+              byteData.containsKey(heightKey) ? toDouble(byteData.get(heightKey)) : null;
+
+          if (width != null || height != null) {
+            int targetWidth = width != null ? toInt(width * density) : bitmap.getWidth();
+            int targetHeight = height != null ? toInt(height * density) : bitmap.getHeight();
+
+            if (width != null && height == null) {
+              // If only width is provided, calculate height based on aspect ratio.
+              double aspectRatio = (double) bitmap.getHeight() / bitmap.getWidth();
+              targetHeight = (int) (targetWidth * aspectRatio);
+            } else if (height != null && width == null) {
+              // If only height is provided, calculate width based on aspect ratio.
+              double aspectRatio = (double) bitmap.getWidth() / bitmap.getHeight();
+              targetWidth = (int) (targetHeight * aspectRatio);
+            }
+            return bitmapDescriptorFactory.fromBitmap(
+                toScaledBitmap(bitmap, targetWidth, targetHeight));
+          } else {
+            // Scale image using given scale ratio
+            final float scale = density / toFloat(byteData.get(imagePixelRatioKey));
+            return bitmapDescriptorFactory.fromBitmap(toScaledBitmap(bitmap, scale));
+          }
+        case "none":
+          break;
+      }
+      return bitmapDescriptorFactory.fromBitmap(bitmap);
+    } catch (Exception e) {
+      throw new IllegalArgumentException("Unable to interpret bytes as a valid image.", e);
+    }
+  }
+
+  /**
+   * Creates a BitmapDescriptor object from asset, using given details and density.
+   *
+   * <p>This method processes an asset specified by name and applies scaling based on the provided
+   * parameters. The `assetDetails` map must contain the keys 'assetName', 'bitmapScaling', and
+   * 'imagePixelRatio', and may optionally include 'width' and/or 'height' to explicitly set the
+   * dimensions of the output image.
+   *
+   * @param assetDetails a map containing the asset details and scaling instructions, with keys
+   *     'assetName': the name of the asset file, 'bitmapScaling': the scaling mode, either 'auto'
+   *     or 'none', 'imagePixelRatio': used with 'auto' scaling to compute the scale ratio, 'width'
+   *     (optional): the desired width, which affects scaling if 'height' is not provided, 'height'
+   *     (optional): the desired height, which affects scaling if 'width' is not provided
+   * @param assetManager assetManager An instance of Android's AssetManager, which provides access
+   *     to any raw asset files stored in the application's assets directory.
+   * @param density density the density of the display, used to calculate pixel dimensions.
+   * @param bitmapDescriptorFactory is an instance of the BitmapDescriptorFactoryWrapper.
+   * @param flutterInjector An instance of the FlutterInjectorWrapper class.
+   * @return BitmapDescriptor object from asset.
+   * @throws IllegalArgumentException if any required keys are missing in `assetDetails` or if the
+   *     asset cannot be opened or processed as a valid image.
+   */
+  @VisibleForTesting
+  public static BitmapDescriptor getBitmapFromAsset(
+      Map<?, ?> assetDetails,
+      AssetManager assetManager,
+      float density,
+      BitmapDescriptorFactoryWrapper bitmapDescriptorFactory,
+      FlutterInjectorWrapper flutterInjector) {
+
+    final String assetNameKey = "assetName";
+    final String bitmapScalingKey = "bitmapScaling";
+    final String imagePixelRatioKey = "imagePixelRatio";
+
+    if (!assetDetails.containsKey(assetNameKey)) {
+      throw new IllegalArgumentException("'asset' requires '" + assetNameKey + "' key.");
+    }
+    if (!assetDetails.containsKey(bitmapScalingKey)) {
+      throw new IllegalArgumentException("'asset' requires '" + bitmapScalingKey + "' key.");
+    }
+    if (!assetDetails.containsKey(imagePixelRatioKey)) {
+      throw new IllegalArgumentException("'asset' requires '" + imagePixelRatioKey + "' key.");
+    }
+
+    final String assetName = toString(assetDetails.get(assetNameKey));
+    final String assetKey = flutterInjector.getLookupKeyForAsset(assetName);
+
+    String scalingMode = toString(assetDetails.get(bitmapScalingKey));
+    switch (scalingMode) {
+      case "auto":
+        final String widthKey = "width";
+        final String heightKey = "height";
+
+        final Double width =
+            assetDetails.containsKey(widthKey) ? toDouble(assetDetails.get(widthKey)) : null;
+        final Double height =
+            assetDetails.containsKey(heightKey) ? toDouble(assetDetails.get(heightKey)) : null;
+        InputStream inputStream = null;
+        try {
+          inputStream = assetManager.open(assetKey);
+          Bitmap bitmap = BitmapFactory.decodeStream(inputStream);
+
+          if (width != null || height != null) {
+            int targetWidth = width != null ? toInt(width * density) : bitmap.getWidth();
+            int targetHeight = height != null ? toInt(height * density) : bitmap.getHeight();
+
+            if (width != null && height == null) {
+              // If only width is provided, calculate height based on aspect ratio.
+              double aspectRatio = (double) bitmap.getHeight() / bitmap.getWidth();
+              targetHeight = (int) (targetWidth * aspectRatio);
+            } else if (height != null && width == null) {
+              // If only height is provided, calculate width based on aspect ratio.
+              double aspectRatio = (double) bitmap.getWidth() / bitmap.getHeight();
+              targetWidth = (int) (targetHeight * aspectRatio);
+            }
+            return bitmapDescriptorFactory.fromBitmap(
+                toScaledBitmap(bitmap, targetWidth, targetHeight));
+          } else {
+            // Scale image using given scale.
+            final float scale = density / toFloat(assetDetails.get(imagePixelRatioKey));
+            return bitmapDescriptorFactory.fromBitmap(toScaledBitmap(bitmap, scale));
+          }
+        } catch (Exception e) {
+          throw new IllegalArgumentException("'asset' cannot open asset: " + assetName, e);
+        } finally {
+          if (inputStream != null) {
+            try {
+              inputStream.close();
+            } catch (IOException e) {
+              e.printStackTrace();
+            }
+          }
+        }
+      case "none":
+        break;
+    }
+
+    return bitmapDescriptorFactory.fromAsset(assetKey);
   }
 
   private static boolean toBoolean(Object o) {
@@ -185,6 +392,13 @@ class Convert {
     return arguments;
   }
 
+  static Messages.PlatformLatLngBounds latLngBoundsToPigeon(LatLngBounds latLngBounds) {
+    return new Messages.PlatformLatLngBounds.Builder()
+        .setNortheast(latLngToPigeon(latLngBounds.northeast))
+        .setSouthwest(latLngToPigeon(latLngBounds.southwest))
+        .build();
+  }
+
   static Object markerIdToJson(String markerId) {
     if (markerId == null) {
       return null;
@@ -239,13 +453,15 @@ class Convert {
     return Arrays.asList(latLng.latitude, latLng.longitude);
   }
 
-  static Object clustersToJson(
-      String clusterManagerId, Set<? extends Cluster<MarkerBuilder>> clusters) {
-    List<Object> data = new ArrayList<>(clusters.size());
-    for (Cluster<MarkerBuilder> cluster : clusters) {
-      data.add(clusterToJson(clusterManagerId, cluster));
-    }
-    return data;
+  static Messages.PlatformLatLng latLngToPigeon(LatLng latLng) {
+    return new Messages.PlatformLatLng.Builder()
+        .setLatitude(latLng.latitude)
+        .setLongitude(latLng.longitude)
+        .build();
+  }
+
+  static LatLng latLngFromPigeon(Messages.PlatformLatLng latLng) {
+    return new LatLng(latLng.getLatitude(), latLng.getLongitude());
   }
 
   static Object clusterToJson(String clusterManagerId, Cluster<MarkerBuilder> cluster) {
@@ -255,7 +471,8 @@ class Convert {
     String[] markerIds = new String[clusterSize];
     MarkerBuilder[] markerBuilders = cluster.getItems().toArray(new MarkerBuilder[clusterSize]);
 
-    // Loops though cluster items and reads markers position for the LatLngBounds builder
+    // Loops though cluster items and reads markers position for the LatLngBounds
+    // builder
     // and also builds list of marker ids on the cluster.
     for (int i = 0; i < clusterSize; i++) {
       MarkerBuilder markerBuilder = markerBuilders[i];
@@ -277,6 +494,27 @@ class Convert {
     return data;
   }
 
+  static Messages.PlatformCluster clusterToPigeon(
+      String clusterManagerId, Cluster<MarkerBuilder> cluster) {
+    int clusterSize = cluster.getSize();
+    String[] markerIds = new String[clusterSize];
+    MarkerBuilder[] markerBuilders = cluster.getItems().toArray(new MarkerBuilder[clusterSize]);
+
+    LatLngBounds.Builder latLngBoundsBuilder = LatLngBounds.builder();
+    for (int i = 0; i < clusterSize; i++) {
+      MarkerBuilder markerBuilder = markerBuilders[i];
+      latLngBoundsBuilder.include(markerBuilder.getPosition());
+      markerIds[i] = markerBuilder.markerId();
+    }
+
+    return new Messages.PlatformCluster.Builder()
+        .setClusterManagerId(clusterManagerId)
+        .setPosition(latLngToPigeon(cluster.getPosition()))
+        .setBounds(latLngBoundsToPigeon(latLngBoundsBuilder.build()))
+        .setMarkerIds(Arrays.asList(markerIds))
+        .build();
+  }
+
   static LatLng toLatLng(Object o) {
     final List<?> data = toList(o);
     return new LatLng(toDouble(data.get(0)), toDouble(data.get(1)));
@@ -295,17 +533,12 @@ class Convert {
         weightedLatLng.getIntensity());
   }
 
-  static Point toPoint(Object o) {
-    Object x = toMap(o).get("x");
-    Object y = toMap(o).get("y");
-    return new Point((int) x, (int) y);
+  static Point pointFromPigeon(Messages.PlatformPoint point) {
+    return new Point(point.getX().intValue(), point.getY().intValue());
   }
 
-  static Map<String, Integer> pointToJson(Point point) {
-    final Map<String, Integer> data = new HashMap<>(2);
-    data.put("x", point.x);
-    data.put("y", point.y);
-    return data;
+  static Messages.PlatformPoint pointToPigeon(Point point) {
+    return new Messages.PlatformPoint.Builder().setX((long) point.x).setY((long) point.y).build();
   }
 
   private static LatLngBounds toLatLngBounds(Object o) {
@@ -352,6 +585,25 @@ class Convert {
     } else {
       return bitmap;
     }
+  }
+
+  private static Bitmap toScaledBitmap(Bitmap bitmap, float scale) {
+    // Threshold to check if scaling is necessary.
+    final float scalingThreshold = 0.001f;
+
+    if (Math.abs(scale - 1) > scalingThreshold && scale > 0) {
+      final int newWidth = (int) (bitmap.getWidth() * scale);
+      final int newHeight = (int) (bitmap.getHeight() * scale);
+      return toScaledBitmap(bitmap, newWidth, newHeight);
+    }
+    return bitmap;
+  }
+
+  private static Bitmap toScaledBitmap(Bitmap bitmap, int width, int height) {
+    if (width > 0 && height > 0 && (bitmap.getWidth() != width || bitmap.getHeight() != height)) {
+      return Bitmap.createScaledBitmap(bitmap, width, height, true);
+    }
+    return bitmap;
   }
 
   private static Point toPoint(Object o, float density) {
@@ -453,7 +705,8 @@ class Convert {
   }
 
   /** Set the options in the given object to marker options sink. */
-  static void interpretMarkerOptions(Object o, MarkerOptionsSink sink) {
+  static void interpretMarkerOptions(
+      Object o, MarkerOptionsSink sink, AssetManager assetManager, float density) {
     final Map<?, ?> data = toMap(o);
     final Object alpha = data.get("alpha");
     if (alpha != null) {
@@ -478,7 +731,7 @@ class Convert {
     }
     final Object icon = data.get("icon");
     if (icon != null) {
-      sink.setIcon(toBitmapDescriptor(icon));
+      sink.setIcon(toBitmapDescriptor(icon, assetManager, density));
     }
 
     final Object infoWindow = data.get("infoWindow");
@@ -564,7 +817,8 @@ class Convert {
     }
   }
 
-  static String interpretPolylineOptions(Object o, PolylineOptionsSink sink) {
+  static String interpretPolylineOptions(
+      Object o, PolylineOptionsSink sink, AssetManager assetManager, float density) {
     final Map<?, ?> data = toMap(o);
     final Object consumeTapEvents = data.get("consumeTapEvents");
     if (consumeTapEvents != null) {
@@ -576,7 +830,7 @@ class Convert {
     }
     final Object endCap = data.get("endCap");
     if (endCap != null) {
-      sink.setEndCap(toCap(endCap));
+      sink.setEndCap(toCap(endCap, assetManager, density));
     }
     final Object geodesic = data.get("geodesic");
     if (geodesic != null) {
@@ -588,7 +842,7 @@ class Convert {
     }
     final Object startCap = data.get("startCap");
     if (startCap != null) {
-      sink.setStartCap(toCap(startCap));
+      sink.setStartCap(toCap(startCap, assetManager, density));
     }
     final Object visible = data.get("visible");
     if (visible != null) {
@@ -790,7 +1044,7 @@ class Convert {
     return pattern;
   }
 
-  private static Cap toCap(Object o) {
+  private static Cap toCap(Object o, AssetManager assetManager, float density) {
     final List<?> data = toList(o);
     switch (toString(data.get(0))) {
       case "buttCap":
@@ -801,9 +1055,10 @@ class Convert {
         return new SquareCap();
       case "customCap":
         if (data.size() == 2) {
-          return new CustomCap(toBitmapDescriptor(data.get(1)));
+          return new CustomCap(toBitmapDescriptor(data.get(1), assetManager, density));
         } else {
-          return new CustomCap(toBitmapDescriptor(data.get(1)), toFloat(data.get(2)));
+          return new CustomCap(
+              toBitmapDescriptor(data.get(1), assetManager, density), toFloat(data.get(2)));
         }
       default:
         throw new IllegalArgumentException("Cannot interpret " + o + " as Cap");
@@ -843,5 +1098,55 @@ class Convert {
       dataArray = (byte[]) data.get("data");
     }
     return new Tile(width, height, dataArray);
+  }
+
+  @VisibleForTesting
+  static class BitmapDescriptorFactoryWrapper {
+    /**
+     * Creates a BitmapDescriptor from the provided asset key using the {@link
+     * BitmapDescriptorFactory}.
+     *
+     * <p>This method is visible for testing purposes only and should never be used outside Convert
+     * class.
+     *
+     * @param assetKey the key of the asset.
+     * @return a new instance of the {@link BitmapDescriptor}.
+     */
+    @VisibleForTesting
+    public BitmapDescriptor fromAsset(String assetKey) {
+      return BitmapDescriptorFactory.fromAsset(assetKey);
+    }
+
+    /**
+     * Creates a BitmapDescriptor from the provided bitmap using the {@link
+     * BitmapDescriptorFactory}.
+     *
+     * <p>This method is visible for testing purposes only and should never be used outside Convert
+     * class.
+     *
+     * @param bitmap the bitmap to convert.
+     * @return a new instance of the {@link BitmapDescriptor}.
+     */
+    @VisibleForTesting
+    public BitmapDescriptor fromBitmap(Bitmap bitmap) {
+      return BitmapDescriptorFactory.fromBitmap(bitmap);
+    }
+  }
+
+  @VisibleForTesting
+  static class FlutterInjectorWrapper {
+    /**
+     * Retrieves the lookup key for a given asset name using the {@link FlutterInjector}.
+     *
+     * <p>This method is visible for testing purposes only and should never be used outside Convert
+     * class.
+     *
+     * @param assetName the name of the asset.
+     * @return the lookup key for the asset.
+     */
+    @VisibleForTesting
+    public String getLookupKeyForAsset(@NonNull String assetName) {
+      return FlutterInjector.instance().flutterLoader().getLookupKeyForAsset(assetName);
+    }
   }
 }