+++ packages/google_maps_flutter/google_maps_flutter_ios/ios/Classes/FLTGoogleMapJSONConversions.m
@@ -165,6 +165,28 @@ NSString *const kHeatmapGradientColorMapSizeKey = @"colorMapSize";
   return nil;
 }
 
++ (NSArray<GMSStrokeStyle *> *)strokeStylesFromPatterns:(NSArray<NSArray<NSObject *> *> *)patterns
+                                            strokeColor:(UIColor *)strokeColor {
+  NSMutableArray *strokeStyles = [[NSMutableArray alloc] initWithCapacity:[patterns count]];
+  for (NSArray *pattern in patterns) {
+    NSString *patternType = pattern[0];
+    UIColor *color = [patternType isEqualToString:@"gap"] ? [UIColor clearColor] : strokeColor;
+    [strokeStyles addObject:[GMSStrokeStyle solidColor:color]];
+  }
+
+  return strokeStyles;
+}
+
++ (NSArray<NSNumber *> *)spanLengthsFromPatterns:(NSArray<NSArray<NSObject *> *> *)patterns {
+  NSMutableArray *lengths = [[NSMutableArray alloc] initWithCapacity:[patterns count]];
+  for (NSArray *pattern in patterns) {
+    NSNumber *length = [pattern count] > 1 ? pattern[1] : @0;
+    [lengths addObject:length];
+  }
+
+  return lengths;
+}
+
 + (GMUWeightedLatLng *)weightedLatLngFromArray:(NSArray<id> *)data {
   NSAssert(data.count == 2, @"WeightedLatLng data must have length of 2");
   if (data.count != 2) {