+++ packages/google_maps_flutter/google_maps_flutter/CHANGELOG.md
@@ -1,6 +1,11 @@
-## 2.7.0
+## 2.8.0
 
 * Adds support for heatmap layers.
+
+## 2.7.0
+
+* Adds support for BitmapDescriptor classes `AssetMapBitmap` and `BytesMapBitmap`.
+* Updates minimum supported SDK version to Flutter 3.22/Dart 3.4.
 
 ## 2.6.1