+++ packages/google_maps_flutter/google_maps_flutter_android/CHANGELOG.md
@@ -1,6 +1,23 @@
-## 2.9.0
+## 2.11.0
 
 * Adds support for heatmap layers.
+
+## 2.10.0
+
+* Converts some platform calls to Pigeon.
+
+## 2.9.1
+
+* Converts inspector interface platform calls to Pigeon.
+
+## 2.9.0
+
+* Adds support for BitmapDescriptor classes `AssetMapBitmap` and `BytesMapBitmap`.
+
+## 2.8.1
+
+* Updates minimum supported SDK version to Flutter 3.22/Dart 3.4.
+* Removes support for apps using the v1 Android embedding.
 
 ## 2.8.0