+++ packages/in_app_purchase/in_app_purchase_storekit/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.3.17
+
+* Removes OCMock from tests.
+
+## 0.3.16
+
+* Converts main plugin class to Swift.
+
 ## 0.3.15
 
 * Replaces `getCountryCode` with `countryCode`.