+++ packages/pointer_interceptor/pointer_interceptor/CHANGELOG.md
@@ -1,8 +1,13 @@
 ## NEXT
 
+* Updates minimum supported SDK version to Flutter 3.16/Dart 3.2.
+
+## 0.10.1+1
+
 * Updates support matrix in README to indicate that iOS 11 is no longer supported.
 * Clients on versions of Flutter that still support iOS 11 can continue to use this
   package with iOS 11, but will not receive any further updates to the iOS implementation.
+* Removes invalid `implements` tag in pubspec.
 
 ## 0.10.1