+++ packages/quick_actions/quick_actions_ios/CHANGELOG.md
@@ -1,5 +1,10 @@
-## NEXT
+## 1.1.1
 
+* Updates to a newer version of Pigeon.
+
+## 1.1.0
+
+* Adds Swift Package Manager compatibility.
 * Updates minimum iOS version to 12.0 and minimum Flutter version to 3.16.6.
 
 ## 1.0.10