+++ packages/video_player/video_player/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 2.8.7
+
+* Ensures that `value.position` never reports a value larger than `value.duration`.
+* Updates minimum supported SDK version to Flutter 3.16/Dart 3.2.
+
 ## 2.8.6
 
 * Updates minimum iOS implementation version to include a privacy manifest.